% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_event_times.R
\name{calculate_event_times}
\alias{calculate_event_times}
\title{This function calculates event times based on an exponential distribution}
\usage{
calculate_event_times(rate, total_time, block_size = 100)
}
\description{
This function calculates event times based on an exponential distribution
}
