% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_cbs_data.R
\name{calculate_transaction_cbs_data}
\alias{calculate_transaction_cbs_data}
\title{This function calculates the cbs data from the transactions}
\usage{
calculate_transaction_cbs_data(tnx_data_tbl, last_date)
}
\description{
This function calculates the cbs data from the transactions
}
