% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnbd_posterior_validation.R
\name{create_pnbd_posterior_validation_data}
\alias{create_pnbd_posterior_validation_data}
\title{This function calculates posterior validation data for the P/NBD model}
\usage{
create_pnbd_posterior_validation_data(
  stanfit,
  data_tbl,
  simparams_tbl,
  bincount = 50
)
}
\description{
This function calculates posterior validation data for the P/NBD model
}
