% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_reparameterisation.R
\name{gamma_mucv2shaperate}
\alias{gamma_mucv2shaperate}
\title{These functions calculates the reparameterising the gamma distributions}
\usage{
gamma_mucv2shaperate(mu, cv)
}
\description{
These functions calculates the reparameterising the gamma distributions
}
