% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_customer_cohort_data.R
\name{generate_customer_cohort_data}
\alias{generate_customer_cohort_data}
\title{This function generates customer cohort data}
\usage{
generate_customer_cohort_data(
  n_customers,
  first_date,
  last_date,
  id_prefix = "C"
)
}
\description{
This function generates customer cohort data
}
