% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pnbd_simulation.R
\name{generate_pnbd_customer_simulation_params}
\alias{generate_pnbd_customer_simulation_params}
\title{This function generates customer simulation parameters based on the P/NBD
model}
\usage{
generate_pnbd_customer_simulation_params(customer_cohort_data_tbl, params_lst)
}
\description{
This function generates customer simulation parameters based on the P/NBD
model
}
