% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{run_pnbd_simulations_chunk}
\alias{run_pnbd_simulations_chunk}
\title{This function runs the simulations for a given set of parameters.}
\usage{
run_pnbd_simulations_chunk(input_param_tbl)
}
\description{
It is expected that a chunk of jobs are combined into a single job.
}
